call custom component in Markdown file
To accomplish that we use MDX that supports react components on markdown.
Then we need `html-to-react` to create a react component from a raw
html.
We also added a LoadExample React component that facilites loading html
files.
All files must be places in components-sources folder which are
loaded during build in docusaurus.config.js.